The Subsetter class is now paremetrized by the source type (i.e. the type of y in y[x]). The idea is to be able to use Subsetter for other things, for example Matrix columns, ...
For lhs use, Subsetter does not materialize. The work is done in the operator=

For rhs use, Subsetter gets to be a sugar expression, which allows us to not only materialize
into a vector, but also to sugar wrap, as in:   sqrt( y[x] )
y[x] is not materialized here.

This is initial rewrite of the feature. For now only dealing with int sugar expressions for the index type.